refactor major and minor key to become part of sealed trait
